requirements.txt
Babel==1.3
coverage==3.7.1
Django==1.8
django-appconf==1.0.1
django-autocomplete-light==2.1.1
django-compressor==1.5
django-debug-toolbar==1.3.0
nodeenv==0.13.1
Pillow==2.8.1
pyjade==3.0.0
pytz==2015.2
six==1.9.0
sqlparse==0.1.15
